refactor: modularize blog service
