+++ apps/cms/src/services/blog/index.ts
@@ -1,0 +1,8 @@
+export { getPosts } from "./posts/list";
+export { getPost } from "./posts/get";
+export { createPost } from "./posts/create";
+export { updatePost } from "./posts/update";
+export { publishPost } from "./posts/publish";
+export { unpublishPost } from "./posts/unpublish";
+export { deletePost } from "./posts/delete";
+export type { SanityPost } from "@platform-core/repositories/blog.server";
