+++ apps/cms/src/services/blog/posts/delete.ts
@@ -1,0 +1,18 @@
+import { deletePost as repoDeletePost } from "@platform-core/repositories/blog.server";
+import { ensureAuthorized } from "../../../actions/common/auth";
+import { getConfig } from "../config";
+
+export async function deletePost(
+  shopId: string,
+  id: string,
+): Promise<{ message?: string; error?: string }> {
+  await ensureAuthorized();
+  const config = await getConfig(shopId);
+  try {
+    await repoDeletePost(config, id);
+    return { message: "Post deleted" };
+  } catch (err) {
+    console.error("Failed to delete post", err);
+    return { error: "Failed to delete post" };
+  }
+}
