+++ apps/cms/src/services/blog/posts/list.ts
@@ -1,0 +1,9 @@
+import { listPosts, type SanityPost } from "@platform-core/repositories/blog.server";
+import { ensureAuthorized } from "../../../actions/common/auth";
+import { getConfig } from "../config";
+
+export async function getPosts(shopId: string): Promise<SanityPost[]> {
+  await ensureAuthorized();
+  const config = await getConfig(shopId);
+  return listPosts(config);
+}
